+++ src/Exception/Fault/AbnormalTerminationException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements ABNORMAL_TERMINATION Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class AbnormalTermination extends Fault
+final class AbnormalTerminationException extends SupervisorException
 {
 }