+++ src/Exception/Fault/AlreadyAddedException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements ALREADY_ADDED Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class AlreadyAdded extends Fault
+final class AlreadyAddedException extends SupervisorException
 {
 }