+++ src/Exception/Fault/AlreadyStartedException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements ALREADY_STARTED Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class AlreadyStarted extends Fault
+final class AlreadyStartedException extends SupervisorException
 {
 }