+++ src/Exception/Fault/BadArgumentsException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements BAD_ARGUMENTS Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class BadArguments extends Fault
+final class BadArgumentsException extends SupervisorException
 {
 }