+++ src/Exception/Fault/BadNameException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements BAD_NAME Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class BadName extends Fault
+final class BadNameException extends SupervisorException
 {
 }