+++ src/Exception/Fault/BadSignalException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements BAD_SIGNAL Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class BadSignal extends Fault
+final class BadSignalException extends SupervisorException
 {
 }