+++ src/Exception/Fault/CantRereadException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements CANT_REREAD Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class CantReread extends Fault
+final class CantRereadException extends SupervisorException
 {
 }