+++ src/Exception/Fault/IncorrectParametersException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements INCORRECT_PARAMETERS Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class IncorrectParameters extends Fault
+final class IncorrectParametersException extends SupervisorException
 {
 }