+++ src/Exception/Fault/NotExecutableException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements NOT_EXECUTABLE Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class NotExecutable extends Fault
+final class NotExecutableException extends SupervisorException
 {
 }