+++ src/Exception/Fault/NotRunningException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements NOT_RUNNING Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class NotRunning extends Fault
+final class NotRunningException extends SupervisorException
 {
 }