+++ src/Exception/Fault/ShutdownStateException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements SHUTDOWN_STATE Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class ShutdownState extends Fault
+final class ShutdownStateException extends SupervisorException
 {
 }