+++ src/Exception/Fault/SignatureUnsupportedException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements SIGNATURE_UNSUPPORTED Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class SignatureUnsupported extends Fault
+final class SignatureUnsupportedException extends SupervisorException
 {
 }