+++ src/Exception/Fault/StillRunningException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements STILL_RUNNING Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class StillRunning extends Fault
+final class StillRunningException extends SupervisorException
 {
 }