+++ src/Exception/Fault/UnknownMethodException.php
@@ -2,13 +2,13 @@
 
 namespace Supervisor\Exception\Fault;
 
-use Supervisor\Exception\Fault;
+use Supervisor\Exception\SupervisorException;
 
 /**
  * Implements UNKNOWN_METHOD Fault exception.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class UnknownMethod extends Fault
+final class UnknownMethodException extends SupervisorException
 {
 }